Rename selectChartOffset to selectChartOffsetInternal and explain this is a private one
