+++ test/state/selectors/selectChartOffset.spec.tsx
@@ -1,14 +1,14 @@
 import React from 'react';
 import { render } from '@testing-library/react';
 import { describe, expect, it } from 'vitest';
-import { selectChartOffset } from '../../../src/state/selectors/selectChartOffset';
+import { selectChartOffsetInternal } from '../../../src/state/selectors/selectChartOffsetInternal';
 import { useAppSelector } from '../../../src/state/hooks';
 import { Customized, LineChart } from '../../../src';
 import { shouldReturnFromInitialState, shouldReturnUndefinedOutOfContext } from '../../helper/selectorTestHelpers';
 
 describe('selectChartOffset', () => {
-  shouldReturnUndefinedOutOfContext(selectChartOffset);
-  shouldReturnFromInitialState(selectChartOffset, {
+  shouldReturnUndefinedOutOfContext(selectChartOffsetInternal);
+  shouldReturnFromInitialState(selectChartOffsetInternal, {
     bottom: 5,
     brushBottom: 5,
     height: 0,
@@ -21,8 +21,8 @@ describe('selectChartOffset', () => {
   it('should be stable', () => {
     expect.assertions(2);
     const Comp = (): null => {
-      const offset1 = useAppSelector(selectChartOffset);
-      const offset2 = useAppSelector(selectChartOffset);
+      const offset1 = useAppSelector(selectChartOffsetInternal);
+      const offset2 = useAppSelector(selectChartOffsetInternal);
       expect(offset1).not.toBe(undefined);
       expect(offset1).toBe(offset2);
       return null;
